requirements.txt
dj-database-url==0.5.0
Django==3.0.3
djangorestframework==3.11.0
gunicorn==20.0.4
whitenoise==5.0.1
django-environ==0.4.5
psycopg2==2.8.4
pytest==5.3.3
